+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetConnection.cpp
@@ -79,6 +79,19 @@ void USpatialNetConnection::UpdateLevelVisibility(const FName& PackageName, bool
 	UpdateActorInterest(Cast<AActor>(PlayerController->GetPawn()));
 }
 
+void USpatialNetConnection::FlushDormancy(AActor* Actor)
+{
+	Super::FlushDormancy(Actor);
+
+	// This gets called from UNetDriver::FlushActorDormancyInternal for each connection. We inject our refresh
+	// of dormancy component here. This is slightly backwards, but means we don't have to make an engine change.
+	if (bReliableSpatialConnection)
+	{
+		const bool bMakeDormant = false;
+		Cast<USpatialNetDriver>(Driver)->RefreshActorDormancy(Actor, bMakeDormant);
+	}
+}
+
 void USpatialNetConnection::UpdateActorInterest(AActor* Actor)
 {
 	if (Actor == nullptr)
@@ -107,7 +120,7 @@ void USpatialNetConnection::ClientNotifyClientHasQuit()
 
 		Worker_ComponentUpdate Update = {};
 		Update.component_id = SpatialConstants::HEARTBEAT_COMPONENT_ID;
-		Update.schema_type = Schema_CreateComponentUpdate(SpatialConstants::HEARTBEAT_COMPONENT_ID);
+		Update.schema_type = Schema_CreateComponentUpdate();
 		Schema_Object* ComponentObject = Schema_GetComponentUpdateFields(Update.schema_type);
 
 		Schema_AddBool(ComponentObject, SpatialConstants::HEARTBEAT_CLIENT_HAS_QUIT_ID, true);
@@ -157,7 +170,7 @@ void USpatialNetConnection::SetHeartbeatEventTimer()
 			Worker_ComponentUpdate ComponentUpdate = {};
 
 			ComponentUpdate.component_id = SpatialConstants::HEARTBEAT_COMPONENT_ID;
-			ComponentUpdate.schema_type = Schema_CreateComponentUpdate(SpatialConstants::HEARTBEAT_COMPONENT_ID);
+			ComponentUpdate.schema_type = Schema_CreateComponentUpdate();
 			Schema_Object* EventsObject = Schema_GetComponentUpdateEvents(ComponentUpdate.schema_type);
 			Schema_AddObject(EventsObject, SpatialConstants::HEARTBEAT_EVENT_ID);