+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetDriver.cpp
@@ -28,6 +28,7 @@
 #include "Interop/SpatialPlayerSpawner.h"
 #include "Interop/SpatialReceiver.h"
 #include "Interop/SpatialSender.h"
+#include "Schema/AlwaysRelevant.h"
 #include "SpatialConstants.h"
 #include "SpatialGDKSettings.h"
 #include "Utils/ActorGroupManager.h"
@@ -38,6 +39,7 @@
 #include "Utils/SpatialMetricsDisplay.h"
 
 #if WITH_EDITOR
+#include "Settings/LevelEditorPlaySettings.h"
 #include "SpatialGDKServicesModule.h"
 #endif
 
@@ -56,6 +58,8 @@ USpatialNetDriver::USpatialNetDriver(const FObjectInitializer& ObjectInitializer
 	, bConnectAsClient(false)
 	, bPersistSpatialConnection(true)
 	, bWaitingForAcceptingPlayersToSpawn(false)
+	, bIsReadyToStart(false)
+	, bMapLoaded(false)
 	, NextRPCIndex(0)
 	, TimeWhenPositionLastUpdated(0.f)
 {
@@ -72,8 +76,6 @@ bool USpatialNetDriver::InitBase(bool bInitAsClient, FNetworkNotify* InNotify, c
 	checkf(!GetReplicationDriver(), TEXT("Replication Driver not supported, please remove it from config"));
 
 	bConnectAsClient = bInitAsClient;
-	bAuthoritativeDestruction = true;
-	bIsReadyToStart = bInitAsClient;
 
 	FCoreUObjectDelegates::PostLoadMapWithWorld.AddUObject(this, &USpatialNetDriver::OnMapLoaded);
 
@@ -159,6 +161,7 @@ bool USpatialNetDriver::InitBase(bool bInitAsClient, FNetworkNotify* InNotify, c
 		}
 	}
 
+	TombstonedEntities.Reserve(EDITOR_TOMBSTONED_ENTITY_TRACKING_RESERVATION_COUNT);
 #endif
 
 	InitiateConnectionToSpatialOS(URL);
@@ -415,6 +418,8 @@ void USpatialNetDriver::OnMapLoaded(UWorld* LoadedWorld)
 			checkNoEntry();
 		}
 	}
+
+	bMapLoaded = true;
 }
 
 void USpatialNetDriver::OnLevelAddedToWorld(ULevel* LoadedLevel, UWorld* OwningWorld)
@@ -655,6 +660,21 @@ void USpatialNetDriver::NotifyActorDestroyed(AActor* ThisActor, bool IsSeamlessT
 			// Remove it from any dormancy lists
 			ClientConnection->DormantReplicatorMap.Remove(ThisActor);
 		}
+
+		// Check if this is a dormant entity, and if so retire the entity
+		if (PackageMap != nullptr)
+		{
+			Worker_EntityId EntityId = PackageMap->GetEntityIdFromObject(ThisActor);
+			if (IsDormantEntity(EntityId) && ThisActor->HasAuthority())
+			{
+				// Deliberately don't unregister the dormant entity, but let it get cleaned up in the entity remove op process
+				if (StaticComponentView->GetAuthority(EntityId, SpatialGDK::Position::ComponentId) != WORKER_AUTHORITY_AUTHORITATIVE)
+				{
+					UE_LOG(LogSpatialOSNetDriver, Warning, TEXT("Retiring dormant entity that we don't have spatial authority over [%lld][%s]"), EntityId, *ThisActor->GetName());
+				}
+				Sender->RetireEntity(EntityId);
+			}
+		}
 	}
 
 	// Remove this actor from the network object list
@@ -676,6 +696,40 @@ void USpatialNetDriver::Shutdown()
 	}
 
 	Super::Shutdown();
+
+	// This is done after Super::Shutdown so the NetDriver is given an opportunity to shutdown all open channels, and those
+	// startup actors that were tombstoned, will be cleaned up also.
+#if WITH_EDITOR
+	const bool bDeleteDynamicEntities = GetDefault<ULevelEditorPlaySettings>()->GetDeleteDynamicEntities();
+
+	if (bDeleteDynamicEntities && IsServer())
+	{
+		for (const Worker_EntityId EntityId : DormantEntities)
+		{
+			if (StaticComponentView->GetAuthority(EntityId, SpatialGDK::Position::ComponentId) == WORKER_AUTHORITY_AUTHORITATIVE)
+			{
+				Connection->SendDeleteEntityRequest(EntityId);
+			}
+		}
+
+		for (const Worker_EntityId EntityId : TombstonedEntities)
+		{
+			if (StaticComponentView->GetAuthority(EntityId, SpatialGDK::Position::ComponentId) == WORKER_AUTHORITY_AUTHORITATIVE)
+			{
+				Connection->SendDeleteEntityRequest(EntityId);
+			}
+		}
+	}
+#endif //WITH_EDITOR
+}
+
+void USpatialNetDriver::NotifyActorFullyDormantForConnection(AActor* Actor, UNetConnection* NetConnection)
+{
+	// Similar to NetDriver::NotifyActorFullyDormantForConnection, however we only care about a single connection
+	const int NumConnections = 1;
+	GetNetworkObjectList().MarkDormant(Actor, NetConnection, NumConnections, this);
+
+	// Intentionally don't call Super::NotifyActorFullyDormantForConnection
 }
 
 void USpatialNetDriver::OnOwnerUpdated(AActor* Actor)
@@ -1419,10 +1473,9 @@ void USpatialNetDriver::TickFlush(float DeltaTime)
 		Sender->FlushPackedRPCs();
 	}
 
-	// Tick the timer manager
-	{
-		TimerManager.Tick(DeltaTime);
-	}
+	ProcessPendingDormancy();
+
+	TimerManager.Tick(DeltaTime);
 
 	Super::TickFlush(DeltaTime);
 }
@@ -1460,6 +1513,9 @@ bool USpatialNetDriver::CreateSpatialNetConnection(const FURL& InUrl, const FUni
 
 	SpatialConnection->InitRemoteConnection(this, nullptr, InUrl, *FromAddr, USOCK_Open);
 	Notify->NotifyAcceptedConnection(SpatialConnection);
+
+	// TODO: This also currently sets all dormant actors to the active list (because the dormancy needs to be processed for the new connection)
+	// This is unnecessary however, as we only have a single relevant connection in Spatial. Could be a performance win to not do this.
 	AddClientConnection(SpatialConnection);
 
 	// Set the unique net ID for this player. This and the code below is adapted from World.cpp:4499
@@ -1500,6 +1556,31 @@ bool USpatialNetDriver::CreateSpatialNetConnection(const FURL& InUrl, const FUni
 	GameMode->GameWelcomePlayer(SpatialConnection, RedirectURL);
 
 	return true;
+}
+
+void USpatialNetDriver::ProcessPendingDormancy()
+{
+	TSet<TWeakObjectPtr<USpatialActorChannel>> RemainingChannels;
+	for (auto& PendingDormantChannel : PendingDormantChannels)
+	{
+		if (PendingDormantChannel.IsValid())
+		{
+			USpatialActorChannel* Channel = PendingDormantChannel.Get();
+			if (Channel->Actor != nullptr)
+			{
+				if (Receiver->IsPendingOpsOnChannel(Channel))
+				{
+					RemainingChannels.Emplace(PendingDormantChannel);
+					continue;
+				}
+			}
+
+			// This same logic is called from within UChannel::ReceivedSequencedBunch when a dormant cmd is received
+			Channel->Dormant = 1;
+			Channel->ConditionalCleanUp(false, EChannelCloseReason::Dormancy);
+		}
+	}
+	PendingDormantChannels = MoveTemp(RemainingChannels);
 }
 
 void USpatialNetDriver::AcceptNewPlayer(const FURL& InUrl, const FUniqueNetIdRepl& UniqueId, const FName& OnlinePlatformName)
@@ -1731,12 +1812,88 @@ USpatialActorChannel* USpatialNetDriver::GetOrCreateSpatialActorChannel(UObject*
 		check(TargetActor);
 		Channel = CreateSpatialActorChannel(TargetActor);
 	}
+#if !UE_BUILD_SHIPPING
+	if (Channel != nullptr && Channel->Actor == nullptr)
+	{
+		// This shouldn't occur, but can often crop up whilst we are refactoring entity/actor/channel lifecycles.
+		UE_LOG(LogSpatialOSNetDriver, Error, TEXT("Failed to correctly initialize SpatialActorChannel for [%s]"), *TargetObject->GetName());
+	}
+#endif // !UE_BUILD_SHIPPING
 	return Channel;
 }
 
 USpatialActorChannel* USpatialNetDriver::GetActorChannelByEntityId(Worker_EntityId EntityId) const
 {
 	return EntityToActorChannel.FindRef(EntityId);
+}
+
+void USpatialNetDriver::RefreshActorDormancy(AActor* Actor, bool bMakeDormant)
+{
+	check(IsServer());
+	check(Actor);
+
+	const Worker_EntityId EntityId = PackageMap->GetEntityIdFromObject(Actor);
+	if (EntityId == SpatialConstants::INVALID_ENTITY_ID)
+	{
+		UE_LOG(LogSpatialOSNetDriver, Verbose, TEXT("Unable to flush dormancy on actor (%s) without entity id"), *Actor->GetName());
+		return;
+	}
+
+	const bool bHasAuthority = StaticComponentView->HasAuthority(EntityId, SpatialConstants::DORMANT_COMPONENT_ID);
+	if (bHasAuthority == false)
+	{
+		UE_LOG(LogSpatialOSNetDriver, Warning, TEXT("Unable to flush dormancy on actor (%s) without authority"), *Actor->GetName());
+		return;
+	}
+
+	const bool bDormancyComponentExists = StaticComponentView->HasComponent(EntityId, SpatialConstants::DORMANT_COMPONENT_ID);
+
+	// If the Actor wants to go dormant, ensure the Dormant component is attached
+	if (bMakeDormant)
+	{
+		if (!bDormancyComponentExists)
+		{
+			Worker_AddComponentOp AddComponentOp{};
+			AddComponentOp.entity_id = EntityId;
+			AddComponentOp.data = SpatialGDK::Dormant().CreateData();
+			Connection->SendAddComponent(AddComponentOp.entity_id, &AddComponentOp.data);
+			StaticComponentView->OnAddComponent(AddComponentOp);
+		}
+	}
+	else
+	{
+		if (bDormancyComponentExists)
+		{
+			Worker_RemoveComponentOp RemoveComponentOp{};
+			RemoveComponentOp.entity_id = EntityId;
+			RemoveComponentOp.component_id = SpatialConstants::DORMANT_COMPONENT_ID;
+			Connection->SendRemoveComponent(EntityId, SpatialConstants::DORMANT_COMPONENT_ID);
+			StaticComponentView->OnRemoveComponent(RemoveComponentOp);
+		}
+	}
+}
+
+void USpatialNetDriver::AddPendingDormantChannel(USpatialActorChannel* Channel)
+{
+	PendingDormantChannels.Emplace(Channel);
+}
+
+void USpatialNetDriver::RegisterDormantEntityId(Worker_EntityId EntityId)
+{
+	// Register dormant entities when their actor channel has been closed, but their entity is still alive.
+	// This allows us to clean them up when shutting down. Might be nice to not rely on ActorChannels to
+	// cleanup in future, but inspect the StaticView and delete all entities that this worker is authoritative over.
+	DormantEntities.Emplace(EntityId);
+}
+
+void USpatialNetDriver::UnregisterDormantEntityId(Worker_EntityId EntityId)
+{
+	DormantEntities.Remove(EntityId);
+}
+
+bool USpatialNetDriver::IsDormantEntity(Worker_EntityId EntityId) const
+{
+	return (DormantEntities.Find(EntityId) != nullptr);
 }
 
 USpatialActorChannel* USpatialNetDriver::CreateSpatialActorChannel(AActor* Actor)
@@ -1886,7 +2043,7 @@ void USpatialNetDriver::DelayedSendDeleteEntityRequest(Worker_EntityId EntityId,
 	FTimerHandle RetryTimer;
 	TimerManager.SetTimer(RetryTimer, [this, EntityId]()
 	{
-		Sender->SendDeleteEntityRequest(EntityId);
+		Sender->RetireEntity(EntityId);
 	}, Delay, false);
 }
 
@@ -1898,18 +2055,28 @@ void USpatialNetDriver::HandleStartupOpQueueing(const TArray<Worker_OpList*>& In
 	}
 
 	QueuedStartupOpLists.Append(InOpLists);
-	bIsReadyToStart = FindAndDispatchStartupOps(InOpLists);
+	if (IsServer())
+	{
+		bIsReadyToStart = FindAndDispatchStartupOpsServer(InOpLists);
+
+		if (bIsReadyToStart)
+		{
+			// We've found and dispatched all ops we need for startup,
+			// trigger BeginPlay() on the GSM and process the queued ops.
+			// Note that FindAndDispatchStartupOps() will have notified the Dispatcher
+			// to skip the startup ops that we've processed already.
+			GlobalStateManager->TriggerBeginPlay();
+		}
+	}
+	else
+	{
+		bIsReadyToStart = FindAndDispatchStartupOpsClient(InOpLists);
+	}
 
 	if (!bIsReadyToStart)
 	{
 	    return;
 	}
-
-	// We've found and dispatched all ops we need for startup, trigger BeginPlay()
-	// on the GSM and process the queued ops.  Note that FindAndDispatchStartupOps()
-	// will have notified the Dispatcher to skip the startup ops that we've
-	// processed already.
-	GlobalStateManager->TriggerBeginPlay();
 
 	for (Worker_OpList* OpList : QueuedStartupOpLists)
 	{
@@ -1924,7 +2091,7 @@ void USpatialNetDriver::HandleStartupOpQueueing(const TArray<Worker_OpList*>& In
 	QueuedStartupOpLists.Empty();
 }
 
-bool USpatialNetDriver::FindAndDispatchStartupOps(const TArray<Worker_OpList*>& InOpLists)
+bool USpatialNetDriver::FindAndDispatchStartupOpsServer(const TArray<Worker_OpList*>& InOpLists)
 {
 	TArray<Worker_Op*> FoundOps;
 
@@ -1970,6 +2137,42 @@ bool USpatialNetDriver::FindAndDispatchStartupOps(const TArray<Worker_OpList*>& 
 		}
 	}
 
+	SelectiveProcessOps(FoundOps);
+
+	if (PackageMap->IsEntityPoolReady() && GlobalStateManager->IsReadyToCallBeginPlay())
+	{
+		// Return whether or not we are ready to start
+		return true;
+	}
+
+	return false;
+}
+
+bool USpatialNetDriver::FindAndDispatchStartupOpsClient(const TArray<Worker_OpList*>& InOpLists)
+{
+	if (bMapLoaded)
+	{
+		return true;
+	}
+	else
+	{
+		// Search for the entity query response for the GlobalStateManager
+		Worker_Op* Op = nullptr;
+		FindFirstOpOfType(InOpLists, WORKER_OP_TYPE_ENTITY_QUERY_RESPONSE, &Op);
+
+		TArray<Worker_Op*> FoundOps;
+		if (Op != nullptr)
+		{
+			FoundOps.Add(Op);
+		}
+
+		SelectiveProcessOps(FoundOps);
+		return false;
+	}
+}
+
+void USpatialNetDriver::SelectiveProcessOps(TArray<Worker_Op*> FoundOps)
+{
 	// For each Op we've found, make a Worker_OpList that just contains that Op,
 	// and pass it to the dispatcher for processing. This allows us to avoid copying
 	// the Ops around and dealing with memory that is / should be managed by the Worker SDK.
@@ -1984,13 +2187,11 @@ bool USpatialNetDriver::FindAndDispatchStartupOps(const TArray<Worker_OpList*>& 
 		Dispatcher->ProcessOps(&SingleOpList);
 		Dispatcher->MarkOpToSkip(Op);
 	}
-
-	if (PackageMap->IsEntityPoolReady() &&
-		GlobalStateManager->IsReadyToCallBeginPlay())
-	{
-		// Return whether or not we are ready to start
-		return true;
-	}
-
-	return false;
 }
+
+#if WITH_EDITOR
+void USpatialNetDriver::TrackTombstone(const Worker_EntityId EntityId)
+{
+	TombstonedEntities.Add(EntityId);
+}
+#endif