+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSettings.h
@@ -18,12 +18,10 @@ struct FWorldLaunchSection
 	FWorldLaunchSection()
 		: Dimensions(2000, 2000)
 		, ChunkEdgeLengthMeters(50)
-		, StreamingQueryIntervalSeconds(4)
 		, SnapshotWritePeriodSeconds(0)
 	{
 		LegacyFlags.Add(TEXT("bridge_qos_max_timeout"), TEXT("0"));
 		LegacyFlags.Add(TEXT("bridge_soft_handover_enabled"), TEXT("false"));
-		LegacyFlags.Add(TEXT("enable_chunk_interest"), TEXT("false"));
 	}
 
 	/** The size of the simulation, in meters, for the auto-generated launch configuration file. */
@@ -33,10 +31,6 @@ struct FWorldLaunchSection
 	/** The size of the grid squares that the world is divided into, in “world units” (an arbitrary unit that worker instances can interpret as they choose). */
 	UPROPERTY(Category = "SpatialGDK", EditAnywhere, config, meta = (ConfigRestartRequired = false, DisplayName = "Chunk edge length in meters"))
 	int32 ChunkEdgeLengthMeters;
-
-	/** The time in seconds between streaming query updates. */
-	UPROPERTY(Category = "SpatialGDK", EditAnywhere, config, meta = (ConfigRestartRequired = false, DisplayName = "Streaming query interval in seconds"))
-	int32 StreamingQueryIntervalSeconds;
 
 	/** The frequency in seconds to write snapshots of the simulated world. */
 	UPROPERTY(Category = "SpatialGDK", EditAnywhere, config, meta = (ConfigRestartRequired = false, DisplayName = "Snapshot write period in seconds"))