+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added several new STAT annotations into the ServerReplicateActors call chain.
 - Avoid generating schema for all UObject subclasses. Actor, ActorComponent, GameplayAbility subclasses are enabled by default, other classes can be enabled using SpatialType UCLASS specifier.
 - Added new experimental CookAndGenerateSchemaCommandlet that generates required schema during a regular cook.
+- Added OverrideSpatialOffloading command line flag that allows toggling of offloading at launch time.
+
+### Breaking Changes:
+- If your project uses replicated subobjects that do not inherit from ActorComponent or GameplayAbility, you need to enable generating schema for them using SpatialType UCLASS specifier or by checking Spatial Type if it's a blueprint.
+- If you already have a project that you are upgrading to this version of the GDK, it is encouraged to follow the upgrade process to SpatialOS `14.1.0`:
+1. Open the `spatialos.json` file in the `spatial/` directory of your project.
+1. Replace the `sdk_version` value and the version value of all dependencies with `14.1.0`.
+1. Replace all other instances of the version number in the file.
 
 ### Features:
 - Visual Studio 2019 is now supported.
@@ -16,8 +24,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a check for schema and snapshot before attempting to start a local deployment. If either are missing then an error message will be displayed.
 - Added optional net relevancy check in replication prioritization. If enabled, an actor will only be replicated if IsNetRelevantFor is true for one of the connected client's views.
 - It is now possible to specify in Unreal which actors should not persist as entities in the Snapshot.
+- Deleted startup actors are now tracked
+- The GDK now uses SpatialOS `14.1.0`.
+- Added a user bindable delegate to SpatialMetrics which triggers when worker metrics have been received.
+- Local deployments now create a new log file known as 'launch.log' which will contain logs relating to starting and running a deployment. Additionally it will contain worker logs which are forwarded to the SpatialOS runtime.
+- Added a new setting to SpatialOS Runtime Settings 'Worker Log Level' which allows configuration of which verbosity of worker logs gets forwarded to the SpatialOS runtime.
+- Added a new developer tool called 'Spatial Output Log' which will show local deployment logs from the 'launch.log' file.
 
 ### Bug fixes:
+- Fixed a bug where the spatial daemon started even with spatial networking disabled.
 - Fixed an issue that could cause multiple Channels to be created for an Actor.
 - PlayerControllers on non-auth servers now have BeginPlay called with correct authority.
 - Attempting to replicate unsupported types (such as TMap) produce a log error rather than crashing the game.
@@ -27,6 +42,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generating schema after deleting the schema database but not the generated schema folder will now correctly trigger an initial schema generation.
 - Streaming levels with QBI enabled no longer produces errors if the player connection owns unreplicated actors.
 - Fixed an issue that would prevent player movement in a zoned deployment.
+- Fixed an issue that could cause queued incoming RPCs with unresolved references to never be processed.
+
+### Breaking Changes:
+- Chunk based interest is no longer supported. All interest is resolved through QBI. You should remove streaming query and chunk based interest options from worker and launch config files to avoid unnecessary streaming queries being generated.
 
 ## [`0.6.1`] - 2019-08-15